playing with a ml api

+++ ml/api.go
@@ -3,6 +3,6 @@ package ml
 // TODO: this is all wrong, I just need a pivot point in the sand
 
 type Classifier interface {
-	Classify(data []float64) (float64, error)
-	Train(data [][]float64, correct []float64) error
+	Classify(data []float64) (int, error)
+	Train(data [][]float64, correct []int) error
 }
